dsmPDFLib/requirements.txt
azure-storage-blob==12.17.0
setuptools >=65
twine >=4.0.2
wheel>=0.42.0